+++ app-backend/batch/src/main/scala/Main.scala
@@ -5,7 +5,8 @@ import com.rasterfoundry.batch.cogMetadata.{HistogramBackfill, OverviewBackfill}
 import com.rasterfoundry.batch.export.{CreateExportDef, UpdateExportStatus}
 import com.rasterfoundry.batch.healthcheck.HealthCheck
 import com.rasterfoundry.batch.aoi.UpdateAOIProject
-import com.rasterfoundry.batch.stac.{ReadStacFeature}
+import com.rasterfoundry.batch.stacImport.{ReadStacFeature}
+import com.rasterfoundry.batch.stacExport.{WriteStacCatalog}
 import com.rasterfoundry.batch.notification.NotifyIngestStatus
 
 object Main {
@@ -17,8 +18,9 @@ object Main {
     NotifyIngestStatus.name -> (NotifyIngestStatus.main(_)),
     OverviewBackfill.name -> (OverviewBackfill.main(_)),
     ReadStacFeature.name -> (ReadStacFeature.main(_)),
+    WriteStacCatalog.name -> (WriteStacCatalog.main(_)),
     UpdateAOIProject.name -> (UpdateAOIProject.main(_)),
-    UpdateExportStatus.name -> (UpdateExportStatus.main(_))
+    UpdateExportStatus.name -> (UpdateExportStatus.main(_)),
   )
 
   def main(args: Array[String]): Unit = {