+++ app-backend/db/src/main/scala/ProjectLayerScenesDao.scala
@@ -17,13 +17,16 @@ object ProjectLayerScenesDao extends Dao[Scene] {
   val tableName =
     "scenes_to_layers s2l INNER JOIN scenes s ON s2l.scene_id = s.id"
   val selectF = fr"""
-      SELECT
+    SELECT
       s.id, s.created_at, s.created_by, s.modified_at, s.owner,
-          s.visibility, s.tags,
-          s.datasource, s.scene_metadata, s.name, s.tile_footprint,
-          s.data_footprint, s.metadata_files, s.ingest_location, s.cloud_cover,
-          s.acquisition_date, s.sun_azimuth, s.sun_elevation, s.thumbnail_status,
-          s.boundary_status, s.ingest_status, s.scene_type FROM""" ++ tableF
+      s.visibility, s.tags,
+      s.datasource, s.scene_metadata, s.name, s.tile_footprint,
+      s.data_footprint, s.metadata_files, s.ingest_location, s.cloud_cover,
+      s.acquisition_date, s.sun_azimuth, s.sun_elevation, s.thumbnail_status,
+      s.boundary_status, s.ingest_status, s.scene_type, s.data_path, s.crs,
+      s.band_count, s.cell_type, s.grid_extent, s.resolutions, s.no_data_value
+    FROM
+  """ ++ tableF
 
   def countLayerScenes(
       projectId: UUID
@@ -42,19 +45,21 @@ object ProjectLayerScenesDao extends Dao[Scene] {
 
   def listLayerScenesRaw(
       layerId: UUID,
-      splitOptions: SplitOptions
-  ): ConnectionIO[List[Scene.ProjectScene]] = {
-    val sceneParams = CombinedSceneQueryParams(
-      sceneParams = SceneQueryParameters(
-        minAcquisitionDatetime = Some(splitOptions.rangeStart),
-        maxAcquisitionDatetime = Some(splitOptions.rangeEnd)
-      )
-    )
+      splitOptionsO: Option[SplitOptions] = None): ConnectionIO[List[Scene]] = {
+    val sceneParams = splitOptionsO match {
+      case Some(splitOptions: SplitOptions) =>
+        CombinedSceneQueryParams(
+          sceneParams = SceneQueryParameters(
+            minAcquisitionDatetime = Some(splitOptions.rangeStart),
+            maxAcquisitionDatetime = Some(splitOptions.rangeEnd)
+          )
+        )
+      case _ => CombinedSceneQueryParams()
+    }
     query
       .filter(fr"project_layer_id = ${layerId}")
       .filter(sceneParams)
       .list
-      .flatMap(scenesToProjectScenes(_, layerId))
   }
 
   def listLayerScenes(
@@ -136,4 +141,19 @@ object ProjectLayerScenesDao extends Dao[Scene] {
     }
   }
 
+  def createUnionedGeomExtent(
+      layerId: UUID
+  ): ConnectionIO[Option[UnionedGeomExtent]] =
+    (fr"""
+    SELECT
+      ST_Transform(ST_Collect(s.data_footprint), 4326) AS geometry,
+      ST_XMin(ST_Extent(ST_Transform(s.data_footprint, 4326))) AS x_min,
+      ST_YMin(ST_Extent(ST_Transform(s.data_footprint, 4326))) AS y_min,
+      ST_XMax(ST_Extent(ST_Transform(s.data_footprint, 4326))) AS x_max,
+      ST_YMax(ST_Extent(ST_Transform(s.data_footprint, 4326))) AS y_max
+    FROM scenes s
+    JOIN scenes_to_layers stl
+    ON s.id = stl.scene_id
+    WHERE stl.project_layer_id = ${layerId}
+  """).query[UnionedGeomExtent].option
 }