+++ app-backend/db/src/main/scala/filters/Filterables.scala
@@ -412,6 +412,20 @@ trait Filterables extends RFMeta with LazyLogging {
     Filterable[Any, TaskQueryParameters] { qp =>
       Filters.taskQP(qp)
     }
+
+  implicit val labelStacExportQPFilter
+    : Filterable[Any, StacExportQueryParameters] =
+    Filterable[Any, StacExportQueryParameters] {
+      params: StacExportQueryParameters =>
+        Filters.onlyUserQP(params.userParams) ++
+          Filters.ownerQP(params.ownerParams) ++
+          Filters.searchQP(params.searchParams, List("name")) ++
+          List(
+            params.exportStatus map { qp =>
+              fr"export_status = UPPER($qp)::public.export_status"
+            },
+          )
+    }
 }
 
 object Filterables extends Filterables